refactor: 인수테스트 페키지 변경과 회원가입 리팩토링
